Move reusable code out of Android.mk backend
The Android.bp backend will share some code with the Android.mk
backend, so move some code out of android_make.go.

Function related to compiler flags have been moved to a new package,
ccflags.go

A new package has been made for WriteIfChanged.
