+++ core/android.go
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2020 Arm Limited.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package core
+
+// Logic common to the Android.mk and Android.bp backends
+
+import (
+	"github.com/google/blueprint"
+)
+
+var (
+	dummyRule = pctx.StaticRule("dummy",
+		blueprint.RuleParams{
+			// We don't want this rule to do anything, so just echo the target
+			Command:     "echo $out",
+			Description: "Dummy rule",
+		})
+)
+
+func enabledAndRequired(m blueprint.Module) bool {
+	if e, ok := m.(enableable); ok {
+		if !isEnabled(e) || !isRequired(e) {
+			return false
+		}
+	}
+	return true
+}
